+++ scripts/graphs.js
@@ -4,27 +4,27 @@ import { Policy } from "@copilot-ld/libpolicy";
 import { ResourceIndex } from "@copilot-ld/libresource";
 import { createStorage } from "@copilot-ld/libstorage";
 import { createLogger } from "@copilot-ld/libutil";
-import { TripleIndex, TripleProcessor } from "@copilot-ld/libtriple";
+import { GraphIndex, GraphProcessor } from "@copilot-ld/libgraph";
 
 /**
- * Processes resources into RDF triples
+ * Processes resources into RDF graphs
  */
 async function main() {
   const resourceStorage = createStorage("resources");
-  const tripleStorage = createStorage("triples");
+  const graphStorage = createStorage("graphs");
   const policyStorage = createStorage("policies");
 
   const policy = new Policy(policyStorage);
   const resourceIndex = new ResourceIndex(resourceStorage, policy);
   const n3Store = new Store();
-  const tripleIndex = new TripleIndex(tripleStorage, n3Store, "triples.jsonl");
-  const logger = createLogger("script.triples");
+  const graphIndex = new GraphIndex(graphStorage, n3Store, "graphs.jsonl");
+  const logger = createLogger("script.graphs");
 
-  const processor = new TripleProcessor(tripleIndex, resourceIndex, logger);
+  const processor = new GraphProcessor(graphIndex, resourceIndex, logger);
 
   const actor = "cld:common.System.root";
 
-  // Process resources into RDF triples (content only)
+  // Process resources into RDF graphs (content only)
   await processor.process(actor);
 }